Hide navigation on mobile and reorganise layout component
