+++ src/components/layout/index.tsx
@@ -1,0 +1,14 @@
+import React from "react"
+import { Box } from "@mui/material"
+import { Header } from "./Header"
+
+const Layout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <Box>
+      <Header />
+      {children}
+    </Box>
+  )
+}
+
+export default Layout
