+++ src/hooks/useDeviceDetect.ts
@@ -1,0 +1,11 @@
+import { useMediaQuery, useTheme } from "@mui/material"
+
+export const useDeviceDetect = () => {
+  const theme = useTheme()
+  const isMobile = useMediaQuery(theme.breakpoints.down("sm"))
+  const isTablet = useMediaQuery(theme.breakpoints.between("sm", "md"))
+  const isLaptop = useMediaQuery(theme.breakpoints.between("md", "lg"))
+  const isDesktop = useMediaQuery(theme.breakpoints.up("lg"))
+
+  return { isMobile, isTablet, isLaptop, isDesktop }
+}
