+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from "next/head"
 import { Box, Container } from "@mui/material"
 import { api } from "@/api/axios"
 import { Book } from "@/api/types"
-import Layout from "@/components/Layout"
+import Layout from "@/components/layout"
 import Banner from "@/components/pages/index/Banner"
 import PopularCategories from "@/components/pages/index/PopularCategories"
 import Top from "../components/pages/index/top"
